fix redux store bug
