+++ client/src/store/modules/newpost.js
@@ -24,7 +24,7 @@ const initialState = {
   location: '',
   story: '',
   landmark: '',
-  profile_id: __PRELOADED_STATE__.currentUser.id,
+  profile_id: __PRELOADED_STATE__.user.user.id,
   allUserStories: [{Test: 'Data'}],
   image_url: '',
 };
